fix a number of mistake due to the refactoring

+++ tests/test_pipeline.sh
@@ -38,7 +38,9 @@ setUp() {
 }
 
 Teardown() {
-	rm -r "$PRD"
+  if [ "$CLEAN_AFTER_TEST" = 1 ]; then
+  	rm -r "$PRD"
+  fi
 }
 
 
@@ -47,24 +49,29 @@ test_function() {
   printf "\n >>> Test "$NAME_TEST" output is : "$out" <<< \n"
 }
 
-#( export NAME_TEST="fsl_5"; export FSL="fsl5.0-"; setUp; test_function ) &
-#( export NAME_TEST="registration_boundary"; export REGISTRATION="boundary"; setUp; test_function; Teardown ) &
-#( export NAME_TEST="registration_pseudo"; export REGISTRATION="pseudo"; setUp; test_function; Teardown ) &
-#( export NAME_TEST="region_mapping_corr"; export REGION_MAPPING_COOR="0.5"; setUp; test_function; Teardown ) &
-#( export NAME_TEST="k_list"; export K_LIST="0 2 5"; setUp; test_function; Teardown  ) &
-#( export NAME_TEST="no_k_list"; export K_LIST=""; setUp; test_function; Teardown ) &
-#( export NAME_TEST="number_tracks"; export NUMBER_TRACKS=10000"; setUp; test_function; Teardown ) &
-( export NAME_TEST="parcel_destrieux"; PARCEL="destrieux"; setUp; test_function ) &
-#( export NAME_TEST="parcel_HCP"; export PARCEL="HCP-MMP"; setUp; test_function ) &
-#( export NAME_TEST="parcel_Yeo7"; export PARCEL="Yeo-7nets"; setUp; test_function) &
-#( export NAME_TEST="parcel_Yeo17"; export PARCEL="Yeo-17nets"; setUp; test_function; Teardown ) &
-#( export NAME_TEST="parcel_no_topup"; export TOPUP="no"; setUp; test_function ) &
-#( export NAME_TEST="no_act"; export ACT="no"; setUp; test_function; Teardown ) &
-#( export NAME_TEST="no_sift"; export SIFT="no""; setUp; test_function; Teardown ) &
-#( export NAME_TEST="sift"; export SIFT="sift""; setUp; test_function; Teardown ) &
-#( export NAME_TEST="sift_multiplier"; export SIFT="sift"; export SIFT_MULTIPLIER=2; setUp; test_function; Teardown ) &
-#( export NAME_TEST="seed_dynamic"; export SEED="dynamic"; setUp; test_function; Teardown ) &
-#( export NAME_TEST="aseg_fs"; export PARCEL="destrieux"; setUp; test_function; Teardown ) &
-#( export NAME_TEST="nb_threads_2"; export ASEG="fs"; setUp; test_function; Teardown ) &
+# Variables to change from config file:
+export TOPUP="no" # to accelerate things a little, eddy takes several hours
+export NB_THREADS=1
+export CLEAN_AFTER_TEST=0
+
+( export NAME_TEST="fsl_5"; export FSL="fsl5.0-"; setUp; test_function ) &
+( export NAME_TEST="registration_boundary"; export REGISTRATION="boundary"; setUp; test_function; Teardown ) &
+( export NAME_TEST="registration_pseudo"; export REGISTRATION="pseudo"; setUp; test_function; Teardown ) &
+( export NAME_TEST="region_mapping_corr"; export REGION_MAPPING_COOR="0.5"; setUp; test_function; Teardown ) &
+( export NAME_TEST="k_list"; export K_LIST="0 2 5"; setUp; test_function; Teardown  ) &
+( export NAME_TEST="no_k_list"; export K_LIST=""; setUp; test_function; Teardown ) &
+( export NAME_TEST="number_tracks"; export NUMBER_TRACKS=10000; setUp; test_function; Teardown ) &
+( export NAME_TEST="parcel_destrieux"; PARCEL="destrieux"; setUp; test_function; Teardown ) &
+( export NAME_TEST="parcel_HCP"; export PARCEL="HCP-MMP"; setUp; test_function; Teardown ) &
+( export NAME_TEST="parcel_Yeo7"; export PARCEL="Yeo-7nets"; setUp; test_function; Teardown) &
+( export NAME_TEST="parcel_Yeo17"; export PARCEL="Yeo-17nets"; setUp; test_function; Teardown ) &
+( export NAME_TEST="no_topup"; export TOPUP="no"; setUp; test_function; Teardown ) &
+( export NAME_TEST="no_act"; export ACT="no"; setUp; test_function; Teardown ) &
+( export NAME_TEST="no_sift"; export SIFT="no""; setUp; test_function; Teardown ) &
+( export NAME_TEST="sift"; export SIFT="sift""; setUp; test_function; Teardown ) &
+( export NAME_TEST="sift_multiplier"; export SIFT="sift"; export SIFT_MULTIPLIER=2; setUp; test_function; Teardown ) &
+( export NAME_TEST="seed_dynamic"; export SEED="dynamic"; setUp; test_function; Teardown ) &
+( export NAME_TEST="aseg_fs"; export PARCEL="destrieux"; setUp; test_function; Teardown ) &
+( export NAME_TEST="nb_threads_2"; export ASEG="fs"; setUp; test_function; Teardown ) &
 
 exit
